Pulled all concrete controllers into Control.py
